LPS-121606 Split code between an interface and an implementation
